+++ wwwsrc/src/components/VaultComponent.vue
@@ -1,6 +1,6 @@
 <template>
     <div id="vault-component">
-        <div style="width: 50vw;" class="card vault-card shadow text-dark">
+        <div style="width: 25vw;" class="card vault-card shadow text-dark">
             <h4 class="p-2">{{vault.name}}</h4>
             <h5 class="p-2">{{vault.description}}</h5>
             <button class="btn btn-primary float-right m-2">View</button>
@@ -15,7 +15,10 @@ export default {
     props: [ "vault" ],
     methods: {
         deleteVault(){
-            this.$store.dispatch("deleteVault", this.vault);
+            let r = confirm("Delete This Vault? \nIt Cannot Be Undone");
+            if(r == true){
+                this .$store.dispatch("deleteVault", this.vault);
+            }
         },
     },
 }
